requirements/base.txt
polars>=0.20,<1.0
pyarrow>=15.0
PyYAML>=6.0
pydantic>=2.6
numpy>=1.26